+++ app/views/admin/article_categories/index.blade.php
@@ -12,13 +12,14 @@
 @section('content')
 <div class="row">
     <div class="col-md-12 center">
-        
+
         <div class="table-wrap">
             <div class="table-header">
-                <a href="{{route('admin.article_categories.create')}}" class="btn btn-sm btn-primary">
-                    <i class="fa fa-plus"></i> <?php echo trans('messages.add_category'); ?>
-                </a>
-               
+                <?php if (in_array('admin.article_categories.create', $allowed_routes)): ?>
+                    <a href="{{route('admin.article_categories.create')}}" class="btn btn-sm btn-primary">
+                        <i class="fa fa-plus"></i> <?php echo trans('messages.add_category'); ?>
+                    </a>
+                <?php endif; ?>
             </div>
             <table class="table table-bordered">
                 <thead>
@@ -30,6 +31,11 @@
                     </tr>
                 </thead>
                 <tbody>
+                    <?php
+                    $allowEdit = in_array('admin.article_categories.edit', $allowed_routes);
+                    $allowDestroy = in_array('admin.article_categories.destroy', $allowed_routes);
+
+                    ?>
                     <?php foreach ($categories as $category): ?>
                         <tr>
                             <td><?php echo $category->id ?></td>
@@ -38,16 +44,21 @@
                             </td>
                             <td><?php echo $category->created_at->format('d/m/Y, H:i') ?></td>
                             <td>
-                                <a href="<?php echo route('admin.article_categories.edit', $category->id) ?>" class="btn btn-xs btn-primary">
-                                    <i class="fa fa-pencil"> <?php echo trans('messages.edit'); ?></i>
-                                </a>
-                                    <?php 
+                                <?php if ($allowEdit): ?>
+                                    <a href="<?php echo route('admin.article_categories.edit', $category->id) ?>" class="btn btn-xs btn-primary">
+                                        <i class="fa fa-pencil"> <?php echo trans('messages.edit'); ?></i>
+                                    </a>
+                                <?php endif; ?>
+                                <?php
+                                if ($allowDestroy):
                                     $deleteUrl = route('admin.article_categories.destroy', $category->id);
-                                    $confirmMsg = trans('confirmation.delete_article_category',array('name' => $category->name ));
+                                    $confirmMsg = trans('confirmation.delete_article_category', array('name' => $category->name));
+
                                     ?>
-                                <a href="<?php echo $deleteUrl ?>" class="btn btn-xs btn-danger" data-method="delete" data-confirm="<?php echo $confirmMsg ?>">
-                                    <i class="fa fa-times"> <?php echo trans('messages.delete'); ?></i>
-                                </a>
+                                    <a href="<?php echo $deleteUrl ?>" class="btn btn-xs btn-danger" data-method="delete" data-confirm="<?php echo $confirmMsg ?>">
+                                        <i class="fa fa-times"> <?php echo trans('messages.delete'); ?></i>
+                                    </a>
+                                <?php endif; ?>
                             </td>
 
                         </tr>
