+++ app/views/admin/product/index.blade.php
@@ -15,9 +15,11 @@
     <div class="col-md-12">
         <div class="table-wrap">
             <div class="table-header">
-                <a href="{{route('admin.product.create')}}" class="btn btn-sm btn-primary">
-                    <i class="fa fa-plus"></i> <?php echo trans('button.new_product'); ?>
-                </a>
+                <?php if (in_array('admin.product.create', $allowed_routes)): ?>
+                    <a href="{{route('admin.product.create')}}" class="btn btn-sm btn-primary">
+                        <i class="fa fa-plus"></i> <?php echo trans('button.new_product'); ?>
+                    </a>
+                <?php endif; ?>
                 <div class="col-md-4 pull-right no-padding">
                     <?php echo View::make('admin.partials.search_tool')->render(); ?>
                 </div>
@@ -33,7 +35,12 @@
                     </tr>
                 </thead>
                 <tbody>
-                    <?php $index = $products->getFrom(); ?>
+                    <?php
+                    $index = $products->getFrom();
+                    $allowEdit = in_array('admin.articles.edit', $allowed_routes);
+                    $allowDestroy = in_array('admin.articles.destroy', $allowed_routes);
+
+                    ?>
                     <?php foreach ($products as $product): ?>
                         <tr>
                             <td><?php echo $index++ ?></td>
@@ -45,16 +52,21 @@
                             <td>{{$product->category->name or ''}}</td>
                             <td><?php echo $product->created_at->format('d/m/Y, H:i') ?></td>
                             <td>
-                                <a href="<?php echo route('admin.articles.edit', $product->id) ?>">
-                                    <i class="fa fa-edit"> <?php echo trans('button.edit'); ?></i>
-                                </a>
+                                <?php if ($allowEdit): ?>
+                                    <a href="<?php echo route('admin.articles.edit', $product->id) ?>">
+                                        <i class="fa fa-edit"> <?php echo trans('button.edit'); ?></i>
+                                    </a>
+                                <?php endif; ?>
                                 <?php
-                                $deleteUrl = route('admin.articles.destroy', $product->id);
-                                $confirmMsg = trans('confirmation.delete_product', array('name' => $product->name));
-                                ?>
-                                <a href="<?php echo $deleteUrl ?>" class="text-danger" data-method="delete" data-confirm="<?php echo $confirmMsg ?>">
-                                    <i class="fa fa-times"> <?php echo trans('button.delete'); ?></i>
-                                </a>
+                                if ($allowDestroy):
+                                    $deleteUrl = route('admin.articles.destroy', $product->id);
+                                    $confirmMsg = trans('confirmation.delete_product', array('name' => $product->name));
+
+                                    ?>
+                                    <a href="<?php echo $deleteUrl ?>" class="text-danger" data-method="delete" data-confirm="<?php echo $confirmMsg ?>">
+                                        <i class="fa fa-times"> <?php echo trans('button.delete'); ?></i>
+                                    </a>
+                                <?php endif; ?>
                             </td>
 
                         </tr>
@@ -69,6 +81,7 @@
                         'to' => $products->getTo(),
                         'total' => $products->getTotal(),
                     ));
+
                     ?>
                 </div>
                 <div class="info-right">
